// src/generated/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format. */
  DateTime: { input: any; output: any; }
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { input: any; output: any; }
};
// ...
export type MutationCreateOneUserArgs = {
  email?: InputMaybe<Scalars['String']['input']>;
  enable_notifications?: InputMaybe<Scalars['Boolean']['input']>;
  instagram?: InputMaybe<Scalars['String']['input']>;
  name?: InputMaybe<Scalars['String']['input']>;
  password?: InputMaybe<Scalars['String']['input']>;
  phone?: InputMaybe<Scalars['String']['input']>;
  picture?: InputMaybe<Scalars['String']['input']>;
  role?: UserRole;
  telegram?: InputMaybe<Scalars['String']['input']>;
  verified_email?: InputMaybe<Scalars['Boolean']['input']>;
};


export type MutationDeleteBonusArgs = {
  id: Scalars['String']['input'];
};


export type MutationDeleteBonusTicketArgs = {
  id: Scalars['String']['input'];
};


export type MutationDeleteOneApplicationArgs = {
  id: Scalars['String']['input'];
};


export type MutationDeleteOneUserArgs = {
  id: Scalars['String']['input'];
};
// ...
export type MutationUpdateMeArgs = {
  email?: InputMaybe<Scalars['String']['input']>;
  enable_notifications?: InputMaybe<Scalars['Boolean']['input']>;
  id: Scalars['String']['input'];
  instagram?: InputMaybe<Scalars['String']['input']>;
  isFullAuth?: InputMaybe<Scalars['Boolean']['input']>;
  name?: InputMaybe<Scalars['String']['input']>;
  password?: InputMaybe<Scalars['String']['input']>;
  phone?: InputMaybe<Scalars['String']['input']>;
  picture?: InputMaybe<Scalars['String']['input']>;
  role?: InputMaybe<UserRole>;
  telegram?: InputMaybe<Scalars['String']['input']>;
  verified_email?: InputMaybe<Scalars['Boolean']['input']>;
};
// ...
export type MutationUpdateOneUserArgs = {
  email?: InputMaybe<Scalars['String']['input']>;
  enable_notifications?: InputMaybe<Scalars['Boolean']['input']>;
  id: Scalars['String']['input'];
  instagram?: InputMaybe<Scalars['String']['input']>;
  isFullAuth?: InputMaybe<Scalars['Boolean']['input']>;
  name?: InputMaybe<Scalars['String']['input']>;
  password?: InputMaybe<Scalars['String']['input']>;
  phone?: InputMaybe<Scalars['String']['input']>;
  picture?: InputMaybe<Scalars['String']['input']>;
  role?: InputMaybe<UserRole>;
  telegram?: InputMaybe<Scalars['String']['input']>;
  verified_email?: InputMaybe<Scalars['Boolean']['input']>;
};
// ...
export type QueryFindOneUserArgs = {
  id: Scalars['String']['input'];
};


export type QueryFindUsersArgs = {
  contact?: InputMaybe<Scalars['String']['input']>;
  id?: InputMaybe<Scalars['String']['input']>;
  name?: InputMaybe<Scalars['String']['input']>;
  role?: InputMaybe<UserRole>;
  skip?: InputMaybe<Scalars['Float']['input']>;
  take?: InputMaybe<Scalars['Float']['input']>;
};
// ...
export type User = {
  __typename?: 'User';
  client_assessment?: Maybe<Scalars['String']['output']>;
  createdAt: Scalars['DateTime']['output'];
  email: Scalars['String']['output'];
  enable_notifications: Scalars['Boolean']['output'];
  google_id?: Maybe<Scalars['String']['output']>;
  id: Scalars['String']['output'];
  instagram?: Maybe<Scalars['String']['output']>;
  isFullAuth: Scalars['Boolean']['output'];
  isGoogleAuth?: Maybe<Scalars['Boolean']['output']>;
  name?: Maybe<Scalars['String']['output']>;
  password?: Maybe<Scalars['String']['output']>;
  phone?: Maybe<Scalars['String']['output']>;
  picture?: Maybe<Scalars['String']['output']>;
  rate?: Maybe<Scalars['Float']['output']>;
  role: UserRole;
  telegram?: Maybe<Scalars['String']['output']>;
  updatedAt: Scalars['DateTime']['output'];
  verified_email?: Maybe<Scalars['Boolean']['output']>;
};

export enum UserRole {
  Admin = 'admin',
  Analyt = 'analyt',
  Blocked = 'blocked',
  Manager = 'manager',
  SuperManager = 'super_manager',
  User = 'user'
}